requirements.txt
fastapi
uvicorn
paddlepaddle
python-dotenv
openai
setuptools
fitz
frontend
PyMuPDF
pydantic